hardware: add 64bit support [1/1]
PD#SH-8930

Problem:
T7 does not have product for android 64 bit

Solution:
Add product t7_an400_arm64 for android 64 bit

Verify:
Verfiy on t7_an400

+++ audio/audio_hal/audio_port.c
@@ -439,7 +439,7 @@ size_t get_inport_consumed_size(input_port *port)
     return port->consumed_bytes;
 }
 
-static ssize_t output_port_start(output_port *port)
+static int output_port_start(output_port *port)
 {
     struct audioCfg cfg = port->cfg;
     struct pcm_config pcm_cfg;
